Split model.py into submodules
